Makes attributes and instance immutable
In order for the struct to be truly immutable it should not be possible to
add methods during runtime and it most importantly it should not be possible
to modify the attributes.
Without this change this is possible by simply:

Example for changing (not reassigning) instance variables
  alice = ImmutableStruct.new(:name).new(name: "alice")
  alice.name.upcase!

Exmaple for changing the object
  instance = ImmutableStruct.new(:name)
  def instance.foo; end

By freezing the object at the end of the constructor and by freezing
all instance variables during initialization this is no longer possible.