Open bundler version requirement in gemspec.
Bundler v2 was released at the beginning of 2019, at which time
CI started pulling down that version, which it turns out is
incompatible with the version we specify in the gemspec.

Travis CI provides a few solutions for how to handle this:
  https://docs.travis-ci.com/user/languages/ruby/#bundler-20

After thinking through the options, I felt like relaxing the
bundler requirement would be least impactful to users of the
gem.

to install bundler - `gem install bundler` - started pulling down
a version of bundler that was incompatible with our gemspec.